fix cell height calculation
The edge of the label is now pinned to the tableview cell and not its
content view. The content views frame is set by the tableview at a
point after this code is run and it cannot be used to properly
determine the cells height. I pin the label to a certain distance from
the cell depending on the cell accessory type.