Temporary solution to the crashing on opening soc link from outside the app.
In the latest version that I have commited to phab , on line 238 and 249 , there is this title . Normally this is filled by the value from the RuFavourite cell in the menu. But when we use open it with a link , the title is nil and we can't add a nil object to a dict it seems , and hence the app craches
the work around is to set the title to an empty string . This way the subject code and course number are shown on top , no title.
I do not know how to solve this properly , as from where do we get the title from ?  May we could do a server request and get it.
But doing a request will be tricky , as we will have to do the request before we show the table , so we will have to do a server request sepearate from the request we make for the cell in the table.